+++ Coffee_iOS/Coffee_iOS/VC/원료커핑/SampleVC.h
@@ -12,7 +12,7 @@
 {
     NSUserDefaults *defaults;
     
-    int box_position;
+    NSInteger box_position;
     NSString *total;
     NSArray *datas;
 }
